ram: fixed int overflow on pagetok macro

+++ components/ram.c
@@ -75,7 +75,7 @@
 	#include <unistd.h>
 
 	#define LOG1024 	10
-	#define pagetok(size, pageshift) ((size) << (pageshift - LOG1024))
+	#define pagetok(size, pageshift) (size_t)(size << (pageshift - LOG1024))
 
 	inline int
 	load_uvmexp(struct uvmexp *uvmexp)
